Add comment "passes" votes; port to jQuery
Refactored all the JS in post.html in order to work with the passes, and
to be honest refactored a lot of main.py as well. I know that I'm going
to have a lot better time if I use jQuery, so rather than hold myself
back, I just did it. Whatever, it barely effects load times.