Update about page: more info, better partitioned
Contains abbreviated links to each element of the philosophy, on the
front page, with more detailed descriptions of each element. I could
take about this stuff for ages, so it's hard to keep it brief and I
think I'm going to revise it, but this is a good starting point, I
think.