Extend verification to /post/<post_id>/link
If a clever scripter looked through the source or even noticed the URL
pattern, they could scrape the site for the post and add /link to the
end. Therefore, you have to check on the post and the link.

I tried to refactor it to a reasonable degree, but it's not quite DRY. I
suppose that's a TODO.