Remove captcha on /submit
I initially started removing the captcha just for editing, and then I
realized there's no purpose having it on post.

Spam is taken care of, because there's a captcha to sign in.

Trolls with one passing post will be downvoted.

If someone makes a script to post repeatedly, I can delete them all and
ban them easily. Then I can add a captcha.

Captchas are bad UX. I /need/ them because of the sensitivity of the
site. But just because I have them implemented, I got a tricky finger
and started using them everywhere. There's no good reason for that. Be
careful.

As of now they're only needed for sign-up and post access for those that
asked for it.